requirements.txt
fastapi
uvicorn[standard]
langchain
langgraph
langchain-google-genai
python-dotenv